feat: implemented continuous waveform shape, 3-band preset, deeper zoom, and robust canvas clipping

+++ src/ui/player/player_state.h
@@ -43,7 +43,7 @@ typedef struct DeckState {
     long long PositionMs;
     long long TrackLengthMs;
 
-    int ZoomScale; // Native Pioneer zoom (1-32)
+    float ZoomScale; // Native Pioneer zoom (1-32, now float for precision)
     float CurrentBPM;
     float OriginalBPM; // Added OriginalBPM missing from struct
